bot/pyproject.toml
[tool.poetry]
name = "bot"
version = "0.1.0"
description = ""
authors = ["RakhmatullinAK"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aiogram = "^2.23.1"
aioredis = "^2.0.1"
python-dotenv = "^0.21.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
